==> Data-processing/0.0  copy Module.ipynb <==
# Copy Necessary File

from libraries.utils import *
import glob

# base folder: For e.g.: 'BusTrails'
base_folder_address = "./Bus_Trails"
up_folder = r'\up'
down_folder = r'\down'
output_folder_name = '.\Trails'

folder_pattern = "\**\DATA*" # ''\**\' is for recursive folders
directories = glob.glob(base_folder_address+folder_pattern)

#Copying

for directory in directories:
    print(f"------ For direcotry: {directory} ------")
    # get path for specific file in the specific directory
    try:
        gps = glob.glob(directory+os.path.sep+'**\*GPS*')[0]
    except IndexError:
        # Skipping trail if GPS file is missing
        print(f"Missing GPS file. Skipping directory {directory}..")
        continue
    
    #extract date from gps file name
    date = os.path.basename(os.path.normpath(gps)).split('_')
    date = date[4]+'-'+date[3]+'-'+date[2]
    
    # making output directory
    output_directory = output_folder_name
    if 'up' in gps:
        output_directory+=up_folder
    elif 'down' in gps:
        output_directory+=down_folder
    output_directory = os.path.join(output_directory, date)
    if os.path.exists(output_directory):
        output_directory+=f"_{os.path.basename(directory)}"
    
    # Fixing and Copying raw GPS file
    print("Reading, fixing and copying raw GPS file...")
    gps_df = read_gps_file(gps)
    save_df(gps_df, output_directory, 'gps_raw')
    
    # Checking existence of WiFi, Sound and ACC files since I need to copy them.
    # copying raw WiFi and Sound files
    try:
        wifi = glob.glob(directory+os.path.sep+'**\*WiFi*')[0]
        print("Copying raw WiFi file...")
        print(f"Copied to: {copyfile(wifi, output_directory)}")
    except IndexError:
        print(f"{directory} missing WiFi file.")
    try:
        sound = glob.glob(directory+os.path.sep+'**\*Sound*')[0]
        print("Copying raw Sound file...")
        print(f"Copied to: {copyfile(sound, output_directory)}")
    except IndexError:
        print(f"{directory} missing Sound file.")
    
    # Fixing and Copying ACC file
    try:
        acc = glob.glob(directory+os.path.sep+'**\*ACC*')[0]
        print("Reading, fixing and copying raw ACC file...")
        acc_df = read_acc_file(acc)
        save_df(acc_df, output_directory, 'acc_raw')
    except IndexError:
        print(f"{directory} missing ACC file.")

    print(f"------------ XXXXXX ------------")
    
print("Completed")

#NICE
