==> Data-processing/0.1  gps Module.ipynb <==
# Analysis_GPS.csv Generator

from libraries.gpsGrouper import *
from libraries.utils import *
import glob

# ouput parent directory of Copy Module for e.g. : '.\Trails'
base_folder_address = "./Trails" 

folder_pattern = "\**\**" # '\**\' is for recursive folders

directories = glob.glob(base_folder_address+folder_pattern)

velocity_threshold = 3
radius_threshold = 30

#Processing

for directory in directories:
    print(f"--------- For directory {directory} ----------")
    # get path for specific file in the specific directory
    gps = glob.glob(directory+os.path.sep+'\GPS.csv')[0]
    
    # process gps
    print("Processing GPS...")
    gps_df = format_gps_df(process_gps(gps, velocity_threshold, radius_threshold))
    
    # saving processed GPS
    save_df(gps_df, directory, 'gps')
    
    print("---------- XXXXXXXXXX ----------")
    
print("Completed")

#NICE
