==> Data-processing/0.4  d_sound_wifi Module.ipynb <==
# Analysis_d.csv Generator.csv

from libraries.dHelper import *
from libraries.wifiGrouper import *
from libraries.soundGrouper import *
from libraries.utils import *
import glob

# ouput parent directory of Copy Module for e.g. : '.\Trails'
base_folder_address = "./Trails" 

folder_pattern = "\**\**" # '\**\' is for recursive folders
processed_gps_filename = 'Analysis_GPS.csv'

directories = glob.glob(base_folder_address+folder_pattern)

d = 20 # in meters
sound_threshold = 70 # in decibels. Keep same value as in soundModule.ipynb

#Processing

for directory in directories:
    print(f"--------- For directory {directory} ----------")
    # Empty df
    d_df = pd.DataFrame()
    try:
        # Load Analysis_GPS
        gps = os.path.join(directory, processed_gps_filename)
        gps_df = pd.read_csv(gps)
        gps_df['start_time'] = gps_df['start_time'].apply(lambda x: time_as_int(x))
        gps_df['end_time'] = gps_df['end_time'].apply(lambda x: time_as_int(x))
        try:
            gps_raw = glob.glob(directory+os.path.sep+'\GPS.csv')[0]
            # For changing start time for d
            print(f"Increasing start_time for d...")
            gps_df = increase_d(d, gps_raw, gps_df)
            gps_df['start_time'] = gps_df['d_start_time']
            gps_df = gps_df[['start_time', 'end_time']]

            # For WiFi with d
            print(f"Processing WiFi for d = {d}...")
            try:
                wifi = glob.glob(directory+os.path.sep+'\*WiFi*.txt')[0]
                wifi_df = process_wifi(wifi, gps_df)
                wifi_df = wifi_df[['wifi_count', 'edge_wifi_count']]
                d_df['d_wifi_count'] = wifi_df['wifi_count']
                d_df['d_edge_wifi_count'] = wifi_df['edge_wifi_count']
            except IndexError: # Missing WiFi file
                print(f"{directory} missing WiFi file. Filling with NaN")
                wifi_df = make_blank_wifi_df(len(gps_df.index))
                d_df['d_wifi_count'] = wifi_df['wifi_count']
                d_df['d_edge_wifi_count'] = wifi_df['edge_wifi_count']
            finally:
                print("Processed WiFi.")
            
            # For sound with d
            print(f"Processing Sound for d = {d}...")
            try:
                sound = glob.glob(directory+os.path.sep+'\*SOUND*.wav')[0]
                sound_df = process_sound(sound, gps_df, sound_threshold)
                sound_df = sound_df[['honk_duration', 'edge_honk_duration']]
                d_df['d_honk_duration'] = sound_df['honk_duration']
                d_df['d_edge_honk_duration'] = sound_df['edge_honk_duration']
            except IndexError: # Missing Sound file
                print(f"{directory} missing Sound file. Filling with NaN")
                sound_df = make_blank_sound_df(len(gps_df.index))
                d_df['d_honk_duration'] = sound_df['honk_duration']
                d_df['d_edge_honk_duration'] = sound_df['edge_honk_duration']
            finally:
                print("Processed Sound.")
            
            # Saving d analysis.
            print(f"Saving analysis for d = {d}...")
            save_df(d_df, directory, 'd-analysis')
            d_df = None
            
        except IndexError:
            print("Raw GPS i.e. GPS.csv not found.")
        except Exception as e:
            raise e
    except FileNotFoundError:
        print(f"Processed GPS i.e. {processed_gps_filename} not found.")
    except Exception as e:
        raise e
    finally:
        print("---------- XXXXXXXXXX ----------")
    
print("Completed")

#NICE
