==> Data-processing/0.5  Sound_MFCC.ipynb <==
# Sound_MFCC.csv Generator

import os
import glob
import librosa
import datetime as d
import pandas as pd
import numpy as np
from scipy.io import wavfile as wav

#Helper functions...

def read_audio(filename):
    sr, _ = wav.read(filename)
    x, sr = librosa.load(filename,sr=sr)
    
    _,_,year,month,day,hour,minute,sec,_=filename.split("\\")[-1].split('_')
    start_time=d.datetime(int(year),int(month),int(day),int(hour),int(minute),int(sec))
    
    time_data=[]
    d_counter=0
    x_len=x.shape[0]

    for i in range(x_len):
        if (i+1)%sr==0:
            time_data.extend([str(start_time)]*sr)
            d_counter+=sr
            start_time+=d.timedelta(seconds=1)
    else:
        remaining=x_len-d_counter
        time_data.extend([str(start_time)]*remaining)
        
    df=pd.DataFrame({'time':time_data,'value':x})
    
    return df,sr


def format_date(mmddyyyy):
    mm,dd,yyyy=mmddyyyy.split('/')
    return '-'.join([yyyy,mm,dd])


def do_mfcc(start_to_end):
    start,end=start_to_end.split('->')
    slot=df_wav[(df_wav.time>=start) & (df_wav.time<=end)].value.values
    mfccs = librosa.feature.mfcc(slot, sr=sr)
    return sorted(list(mfccs.mean(axis=1)),reverse=True)[:MFCC_COMPONENT]

#Hyper-Parameters...

MFCC_COMPONENT=5
folders=glob.glob("./Trails/*/*/")
num_folders=len(folders)

folder=folders[0]
df_wav=read_audio(glob.glob(folder+'*.wav')[0])
df_ana_gps=pd.read_csv(glob.glob(folder+'Analysis_GPS.csv')[0])

df_time_slots=\
pd.DataFrame(df_ana_gps.start_date.apply(format_date)+' '+\
df_ana_gps.start_time+'->'+\
df_ana_gps.end_date.apply(format_date)+' '+\
df_ana_gps.end_time,columns=['time_slots'])

res=df_time_slots.time_slots.apply(do_mfcc)
df_mfcc=pd.DataFrame(list(res),columns=[f'mfcc{i}' for i in range(MFCC_COMPONENT)])
df_mfcc.head()

#Processing....

for i,folder in enumerate(folders):
    if os.path.isfile(folder+'Sound_MFCC.csv'): #checking weither Sound_MFCC file exist already?
        print(f'{i+1}/{num_folders}. folder: {folder} is already processed and Augumented\n') #file is already processed
    else:
        df_wav,sr=read_audio(glob.glob(folder+'*.wav')[0])
        df_ana_gps=pd.read_csv(folder+'Analysis_GPS.csv')
        
        df_time_slots=\
                pd.DataFrame(df_ana_gps.start_date.apply(format_date)+' '+\
                df_ana_gps.start_time+'->'+\
                df_ana_gps.end_date.apply(format_date)+' '+\
                df_ana_gps.end_time,columns=['time_slots'])
        
        res=df_time_slots.time_slots.apply(do_mfcc)
        
        df_mfcc=pd.DataFrame(list(res),columns=[f'mfcc{i}' for i in range(MFCC_COMPONENT)])
        df_mfcc.to_csv(folder+'Sound_MFCC.csv',index=False)

        print(f'{i+1}/{num_folders}. Processed "Analysis_GPS.csv" and added file "Sound_MFCC.csv" in folder: {folder}\n')
        
print('Completed.')

#NICE
