==> Data-processing/1.0  RSI calculation Module.ipynb <==
# RSI.csv Generator

import os
import glob
import numpy as np
import pandas as pd
from math import cos, asin, sqrt, pi

def distance(lat1, lon1, lat2, lon2):
    p = pi/180
    a = 0.5 - cos((lat2-lat1)*p)/2 + cos(lat1*p) * cos(lat2*p) * (1-cos((lon2-lon1)*p))/2
    return 12742000 * asin(sqrt(a))

# Helpers

def orient_zaxis(x,y,z):
    g=9.8
    if(abs(abs(z)-g)<abs(abs(y)-g) and abs(abs(z)-g)<abs(abs(x)-g)):
        #print("Z Normal")
        return z
    elif(abs(abs(x)-g)<abs(abs(z)-g) and abs(abs(x)-g)<abs(abs(y)-g)):
        #print("X=Z")
        return x
    elif(abs(abs(y)-g)<abs(abs(z)-g) and abs(abs(y)-g)<abs(abs(x)-g)):
        #print("Y=Z")
        return y 
    else:
        return z

def get_time_and_vel_in_distance(stop_lat,stop_long,stop_time,vel_df,d):
    for lat,long,time in vel_df[['#lat','long','time']].values:
        #print(lat,long,time)
        if distance(stop_lat,stop_long,lat,long)<=d:
            vel=vel_df[np.logical_and(vel_df.time>=time,vel_df.time<=stop_time)].speed
            return time,stop_time,vel

        
def get_oriented_z_accleration_data_on_timedelta(time,stop_time,acc_df):
    z_acc_oriented=\
    acc_df[np.logical_and(acc_df.time>=time, acc_df.time<=stop_time)].apply(lambda tup:orient_zaxis(tup[0],
                                                                                            tup[1],
                                                                                            tup[2]),
                                                                    axis=1)
    return z_acc_oriented


def get_RSI(z_acc,vel):
    vel_sum=vel.sum()
    if vel_sum>0:
        return (vel.size*np.sqrt(np.square(z_acc).mean()))/vel_sum
    else:
        return 0


def get_rsi_from_d_meter_behind_to_lat_long_time(stop_lat,stop_long,stop_time,vel_df,acc_df,d):
    time,stop_time,vel=get_time_and_vel_in_distance(stop_lat,stop_long,stop_time,vel_df,d)
    z_acc=get_oriented_z_accleration_data_on_timedelta(time,stop_time,acc_df)
    rsi=get_RSI(z_acc,vel)
    return rsi

# Processing RSI

folders=glob.glob('./Trails/*/*/')
road_distance=50

for folder in folders:
    if os.path.isfile(folder+'RSI.csv'): #checking weither RSI file exist already?
        print(f'folder: {folder} is already processed and Augumented\n') #file is already processed
    else:
        acc_df=pd.read_csv(folder+'ACC.csv')
        ana_df=pd.read_csv(folder+'Analysis_GPS.csv')
        vel_df=pd.read_csv(folder+'GPS.csv')

        ana_df['RSI']=\
        ana_df[['lat',
                'long',
                'start_date',
                'start_time']].apply(lambda t:get_rsi_from_d_meter_behind_to_lat_long_time(t[0],
                                                                                           t[1],
                                                                                           t[2]+" "+t[3],
                                                                                           vel_df,
                                                                                           acc_df,
                                                                                           road_distance),
                                     axis=1)

        ana_df[['RSI']].to_csv(folder+'RSI.csv',index=False)
        print(f'Processed and added RSI to file "RSI.csv" in folder: {folder}\n')

#NICE
