==> Data-processing/1.1  Map Imaging Module.ipynb <==
# Map_POI.csv Generator

%matplotlib inline
import os
import glob
import numpy as np
import pandas as pd
from PIL import Image
import matplotlib.pyplot as plt
from libraries.mapImaging import Get_Style_Map_Image_Between,style

#Constants

width,height=300,300

meter_to_degrees=lambda m:(0.001/111)*m

deg_width_half_m=meter_to_degrees(width/2)
deg_height_half_m=meter_to_degrees(height/2)


POI={
    'human_made':  (255,235,59),
    'natural_land':(219,136,54),
    'high_way':    (219,54,54),
    'two_way':     (70,54,219),
    'one_way':     (244,75,137),
    'water':       (33,208,224),
    'park':        (0,255,0),
    'school':      (255,255,255),
    'medical':     (255,0,0),
    'other_poi':   (128,128,128)
}

def get_image_from_center_w_and_h(center_cord,style):

    top_left_cord=(center_cord[0]+deg_height_half_m,center_cord[1]-deg_width_half_m)
    bottom_right_cord=(center_cord[0]-deg_height_half_m,center_cord[1]+deg_width_half_m)

    image=Get_Style_Map_Image_Between(top_left_cord,
                                      bottom_right_cord,
                                      style)
    #plt.imshow(image)
    
    return image
#............................................................................................#

def get_poi_percentage(image,poi,d=2):
    """
    calculate a perticular POI percentage in a image
    """
    r,g,b=POI[poi]
    target=np.abs(image-np.array([r,g,b]))<=d 
    #API gives +1,-1 difference than encoding so introduced a tollerence of d

    poi_map=np.logical_and(np.logical_and(target[:,:,0],target[:,:,1]),target[:,:,2])

    #fig=plt.figure()
    #ax=fig.add_subplot(111)
    #ax.imshow(poi_map)

    return poi_map.mean()
#.............................................................................................#

def get_poi_feat(image):
    """
    calculate all POI percentage and return a dict of all POIs and their percentages
    """
    poi_feat={}

    for poi in POI.keys():
        poi_feat[poi]=get_poi_percentage(image,poi)
        
    total=sum(poi_feat.values()) #total value
    for k in poi_feat.keys():
        poi_feat[k]/=total #normalising

    return poi_feat

#File name structure creator#####################################################################
format_name=lambda fn:fn[2:].replace('\\','_')

#Final function##################################################################################
def Map_POI_surround(index,center_cord,path):
    image=get_image_from_center_w_and_h(center_cord,style)
    img=Image.fromarray(image)
    img.save(path+'images/'+format_name(path)+str(int(index))+'.png')
    poi_feat=get_poi_feat(image)
    return poi_feat

#MKDIR###########################################################################################
def mkdir(path):
    try:
        os.mkdir(path)
        print('Successfully made: '+path)
    except:
        print("Allready exist: "+path)

# Data Aquisation

folders=glob.glob("./Trails/*/*/")

for folder in folders:
    if os.path.isfile(folder+'Map_POI.csv'): #checking weither Map_POI file exist already?
        print(f'folder: {folder} is already processed and Augumented\n') #file is already processed
    else:
        mkdir(folder+'images') #creating folder for image
        df=pd.read_csv(folder+'Analysis_GPS.csv') #reading data
        df=df[['lat','long']].reset_index()
        poi=df[['index','lat','long']].apply(lambda e:Map_POI_surround(e[0],(e[1],e[2]),folder),axis=1) 
        #getting poi data & images
        map_df=pd.DataFrame(list(poi))#poi dataframe
        map_df.to_csv(folder+'Map_POI.csv',index=False) #saving to a new file

        print(f'Processed "Analysis_GPS.csv" and added Map features to file "Map_POI.csv" in folder: {folder}\n')

#NICE
