==> Data-processing/2.0  Static stopage labeling module.ipynb <==
# static_stapages.csv Generator

import glob
import numpy as np
import pandas as pd
from math import cos, asin, sqrt, pi

def distance(lat1, lon1, lat2, lon2):
    p = pi/180
    a = 0.5 - cos((lat2-lat1)*p)/2 + cos(lat1*p) * cos(lat2*p) * (1-cos((lon2-lon1)*p))/2
    return 12742000 * asin(sqrt(a))

# Helpers

def if_in_range_of_tag_df(lat1,long1,dis,tag_df,tag_name):
    for lat,long in tag_df.values:
        if distance(lat1,long1,lat,long)<=dis:
            return tag_name
    else:
        return 'Not '+tag_name

    
def label_dataframe_with_tags(df,tags,stopage_distance=50):
    label={}
    for tag_name in tags.keys():
        l=[]
        for lat,long in df[['lat','long']].values:
            l.append(if_in_range_of_tag_df(lat,long,stopage_distance,tags[tag_name],tag_name))
        label['Is_'+tag_name]=l
        
    return pd.DataFrame(label)

# Tagging

stopage_distance=30 #tollerable distance from the ground truth

for direction in ['down','up']:
    tag_folder=f'./Labels/{direction}/'
    tags=\
        dict(
        [
            ('Bus_stop',pd.read_csv(tag_folder+f'Bus_stop_{direction}.csv',header=None)),
            ('Turn',pd.read_csv(tag_folder+f'Turn_{direction}.csv',header=None)),
            ('Signal',pd.read_csv(tag_folder+f'Signal_{direction}.csv',header=None))

        ])    
    

    folders=glob.glob(f'./Trails/{direction}/*/')
    for folder in folders:
        df=pd.read_csv(folder+'Analysis_GPS.csv')
        label=label_dataframe_with_tags(df,tags,stopage_distance)
        label.to_csv(folder+'static_stopages.csv',index=False) 
        print(f'Labeled trail at folder: {folder}')

#NICE
