==> Data-processing/2.2  adhoc stopage labelling module.ipynb <==
# adhoc_stopages.csv Generator

import glob
import numpy as np
import pandas as pd

# Helpers

def if_adhoc(tup):
    if 'Not' in tup[0] and 'Not' in tup[1] and 'Not' in tup[2] and 'Not' in tup[3]:
        return 'Adhoc'
    else:
        return 'Not Adhoc'

# Tagging

folders=glob.glob('./Trails/*/*/')

for folder in folders:

    stat_lbl_df=pd.read_csv(folder+'static_stopages.csv')
    dyn_lbl_df=pd.read_csv(folder+'dynamic_stopages.csv')

    df=pd.concat([stat_lbl_df,dyn_lbl_df],axis=1)

    df['Is_Adhoc']=df.apply(if_adhoc,axis=1)

    df[['Is_Adhoc']].to_csv(folder+'adhoc_stopages.csv',index=False)
    print(f'Labeled trail at folder: {folder}')

#NICE
