==> Data-processing/3.0  Concatinate data per trail and saving whole module.ipynb <==
# ALL_DATA.csv Generator

import glob
import numpy as np
import pandas as pd

# Concatinating all data

files=['Analysis_GPS.csv',
       'Analysis_WiFi.csv',
       'Analysis_Sound.csv',
       'Analysis_d.csv',
       'Sound_MFCC.csv',
       'Map_POI.csv',
       'RSI.csv',
       'static_stopages.csv',
       'dynamic_stopages.csv',
       'adhoc_stopages.csv']

folders=glob.glob('./Trails/*/*/')

for folder in folders:
    df=pd.concat([pd.read_csv(folder+file) for file in files],axis=1)
    df.to_csv(folder+'All_DATA.csv',index=False)
    print(f'Processed folder: {folder}')

# Saving Trails to 1 file

format_name=lambda fn:fn[2:].replace('\\','_') #this format is used in MapImaging module

all_folders=glob.glob('./Trails/*/*/')

def process_ALL_DATA_for_image_paths(folder):
    df=pd.read_csv(folder+'ALL_DATA.csv').reset_index()
    df['image_path']=df['index'].apply(lambda e:folder.replace("\\","/")+'images/'+format_name(folder)+str(int(e))+'.png')
    df=df.drop(columns=['index'])
    if df.shape[1]!=32+5:
        return None
    return df

whole_df=pd.concat([process_ALL_DATA_for_image_paths(folder) for folder in all_folders],axis=0)

whole_df.head()

whole_df.to_csv('Processed_Bus_Trail_data.csv',index=False)

#NICE
