==> Data-processing/3.3 data_length.ipynb <==
import glob
import numpy as np
import pandas as pd
from tqdm import tqdm
from geopy.distance import geodesic

gps_files=glob.glob("./Bus_Trails/*/*/*/*_GPS_*")

def get_length_of_route(gps_file):
    df=pd.read_csv(gps_file).reset_index()
    df.columns=['lat','long','speed','aultitude','time','remark']

    dis=0
    start_lat,start_long=df[['lat','long']].iloc[0]
    for lat,long in df[['lat','long']].values:
        dis+=geodesic((start_lat,start_long),(lat,long)).m
        start_lat,start_long=lat,long
        
    #print(gps_file,dis)
    return dis

dis=0
for trail in tqdm(gps_files):
    dis+=get_length_of_route(trail)

dis/1000

#NICE
